Fix merge issues for validate_matchdict. Split validate_wtforms_schema to ext.wtforms.
